🚚 server: rename keeper module to wallet
