+++ server/test/utils/wallet.test.ts
@@ -1,6 +1,6 @@
 import "../mocks/deployments";
-import { keeperClient, nonceSource } from "../mocks/keeper";
 import "../mocks/sentry";
+import { nonceSource, walletClient } from "../mocks/wallet";
 
 import { captureException, withScope } from "@sentry/node";
 import { setImmediate } from "node:timers/promises";
@@ -10,9 +10,9 @@ import { afterEach, describe, expect, inject, it, vi } from "vitest";
 import { dataSuffix } from "@exactly/common/attribution";
 import { auditorAbi } from "@exactly/common/generated/chain";
 
-import keeper from "../../utils/keeper";
 import nonceManager from "../../utils/nonceManager";
 import publicClient from "../../utils/publicClient";
+import keeper from "../../utils/wallet";
 
 import type * as sentry from "@sentry/node";
 import type * as timers from "node:timers/promises";
@@ -79,9 +79,9 @@ describe("fault tolerance", () => {
     const waitForTransactionReceipt = publicClient.waitForTransactionReceipt;
     const hardReset = vi.spyOn(nonceManager, "hardReset");
     const currentNonce = await nonceSource.get({
-      address: keeperClient.account.address,
-      chainId: keeperClient.chain.id,
-      client: keeperClient,
+      address: walletClient.account.address,
+      chainId: walletClient.chain.id,
+      client: walletClient,
     });
     const getNonce = vi.spyOn(nonceSource, "get");
     getNonce
@@ -138,9 +138,9 @@ describe("fault tolerance", () => {
     const waitForTransactionReceipt = publicClient.waitForTransactionReceipt;
     const hardReset = vi.spyOn(nonceManager, "hardReset");
     const currentNonce = await nonceSource.get({
-      address: keeperClient.account.address,
-      chainId: keeperClient.chain.id,
-      client: keeperClient,
+      address: walletClient.account.address,
+      chainId: walletClient.chain.id,
+      client: walletClient,
     });
 
     const getNonce = vi.spyOn(nonceSource, "get");
@@ -199,9 +199,9 @@ describe("fault tolerance", () => {
     await vi.waitUntil(
       async () =>
         (await nonceSource.get({
-          address: keeperClient.account.address,
-          chainId: keeperClient.chain.id,
-          client: keeperClient,
+          address: walletClient.account.address,
+          chainId: walletClient.chain.id,
+          client: walletClient,
         })) ===
         currentNonce + 102,
     );
